refactor hierarchy: better division
