+++ alchemist/alchemist-interfaces/src/main/kotlin/it/unibo/alchemist/model/interfaces/EnvironmentWithObstacles.kt
@@ -12,9 +12,9 @@ package it.unibo.alchemist.model.interfaces
 /**
  * An environment with obstacles.
  *
- * @param T the concentration type.
- * @param P the position and vector type for this environment.
  * @param W the type of obstacles.
+ * @param T the concentration type.
+ * @param P the position type for this environment.
  */
 interface EnvironmentWithObstacles<W : Obstacle<P>, T, P : Position<P>> : Environment<T, P> {
 
@@ -30,7 +30,7 @@ interface EnvironmentWithObstacles<W : Obstacle<P>, T, P : Position<P>> : Enviro
      * Removes an obstacle from this environment.
      *
      * @param obstacle
-     *              the obstacle to add
+     *              the obstacle to remove
      * @return true if the obstacle has actually been removed
      */
     fun removeObstacle(obstacle: W): Boolean
@@ -39,23 +39,6 @@ interface EnvironmentWithObstacles<W : Obstacle<P>, T, P : Position<P>> : Enviro
      * @return a list of all the obstacles in this environment.
      */
     fun getObstacles(): List<W>
-
-    /**
-     * Given a point and a range, retrieves all the obstacles within.
-     *
-     * @param center
-     *              the center point
-     * @param range
-     *              the range to scan
-     * @return the list of obstacles
-     */
-    fun getObstaclesInRange(center: P, range: Double): List<W>
-
-    /**
-     * @return true if this environment has mobile obstacles, false if
-     * the obstacles are static
-     */
-    fun hasMobileObstacles(): Boolean
 
     /**
      * Checks whether there is at least an obstacle intersecting the line connecting [start] and [end].
@@ -67,21 +50,4 @@ interface EnvironmentWithObstacles<W : Obstacle<P>, T, P : Position<P>> : Enviro
      * @return true if the line connecting start and end touches an obstacle
      */
     fun intersectsObstacle(start: P, end: P): Boolean
-
-    /**
-     * This method must calculate the ABSOLUTE next allowed position given the
-     * current position and the position in which the node wants to move. For
-     * example, if your node is in position [2,3], wants to move to [3,4] but
-     * the next allowed position (because, e.g., of physical obstacles) is
-     * [2.5,3.5], the result must be a Position containing coordinates
-     * [2.5,3.5].
-     *
-     * @param current
-     *              the current position
-     * @param desired
-     *              the desired position
-     * @return the next allowed position, where the node can actually move. This
-     * position MUST be considered as a vector whose start point is [current].
-     */
-    fun next(current: P, desired: P): P
 }
