+++ alchemist/alchemist-interfaces/src/main/kotlin/it/unibo/alchemist/model/interfaces/Obstacle.kt
@@ -19,22 +19,7 @@ import java.io.Serializable
 interface Obstacle<P : Position<P>> : Serializable {
 
     /**
-     * @return the id for this obstacle
+     * The id for this obstacle.
      */
-    fun getId(): Int
-
-    /**
-     * Given a vector (starting point and end point) representing a requested
-     * move, this method computes a new end point, representing a cut version of
-     * the initial vector, modified in such a way that the end point is outside
-     * the obstacle.
-     *
-     * @param start
-     *          starting point of the vector
-     * @param end
-     *          ending point of the vector
-     * @return the intersection point between the vector and the obstacle nearest
-     *         to the vector's starting point.
-     */
-    fun next(start: P, end: P): P
+    val id: Int
 }
